// src/_proto/Protos/reviews.ts
export interface SubmitReviewVoteRequest {
    /**
     * @generated from protobuf field: int32 id = 1;
     */
    id: number;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReviewType type = 2;
     */
    type: ReviewType;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReviewItem.VoteType voteType = 3;
     */
    voteType: ReviewItem_VoteType;
}
// ...
export interface ReviewerSeriesInfo {
    /**
     * @generated from protobuf field: bool isChampion = 1;
     */
    isChampion: boolean;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.ReviewerEntityInfo
 */
export interface ReviewerEntityInfo {
    /**
     * @generated from protobuf oneof: entityInfo
     */
    entityInfo: {
        oneofKind: "seriesInfo";
        /**
         * @generated from protobuf field: wuxiaworld.api.v2.ReviewerSeriesInfo seriesInfo = 1;
         */
        seriesInfo: ReviewerSeriesInfo;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.ReviewItem
 */
export interface ReviewItem {
    /**
     * @generated from protobuf field: int32 id = 1;
     */
    id: number;
    /**
     * @generated from protobuf field: google.protobuf.StringValue content = 2;
     */
    content?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp reviewedAt = 3;
     */
    reviewedAt?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp editedAt = 4;
     */
    editedAt?: Timestamp;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.UserItem reviewUser = 5;
     */
    reviewUser?: UserItem;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReviewItem.PastUser reviewUserPast = 6;
     */
    reviewUserPast?: ReviewItem_PastUser;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReviewItem.Votes votes = 7;
     */
    votes?: ReviewItem_Votes;
    /**
     * @generated from protobuf field: int32 totalComments = 8;
     */
    totalComments: number;
    /**
     * @generated from protobuf oneof: type
     */
    type: {
        oneofKind: "novelId";
        /**
         * @generated from protobuf field: int32 novelId = 9;
         */
        novelId: number;
    } | {
        oneofKind: "productId";
        /**
         * @generated from protobuf field: int32 productId = 10;
         */
        productId: number;
    } | {
        oneofKind: undefined;
    };
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReviewItem.VoteType voteType = 11;
     */
    voteType: ReviewItem_VoteType;
    /**
     * @generated from protobuf field: bool hasPurchased = 12;
     */
    hasPurchased: boolean;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReviewItem.RejectionInfo rejectionInfo = 13;
     */
    rejectionInfo?: ReviewItem_RejectionInfo;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReviewerEntityInfo reviewerEntityInfo = 14;
     */
    reviewerEntityInfo?: ReviewerEntityInfo;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.ReviewItem.Votes
 */
export interface ReviewItem_Votes {
    /**
     * @generated from protobuf field: int32 upVotes = 1;
     */
    upVotes: number;
    /**
     * @generated from protobuf field: int32 downVotes = 2;
     */
    downVotes: number;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.ReviewItem.RejectionInfo
 */
export interface ReviewItem_RejectionInfo {
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReviewItem.RejectionInfo.State state = 1;
     */
    state: ReviewItem_RejectionInfo_State;
    /**
     * @generated from protobuf field: google.protobuf.StringValue reason = 2;
     */
    reason?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp rejectedAt = 3;
     */
    rejectedAt?: Timestamp;
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.ReviewItem.RejectionInfo.State
 */
export enum ReviewItem_RejectionInfo_State {
    /**
     * @generated from protobuf enum value: Default = 0;
     */
    Default = 0,
    /**
     * @generated from protobuf enum value: Rejected = 1;
     */
    Rejected = 1,
    /**
     * @generated from protobuf enum value: ReApproval = 2;
     */
    ReApproval = 2,
    /**
     * @generated from protobuf enum value: Approved = 3;
     */
    Approved = 3
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.ReviewItem.PastUser
 */
export interface ReviewItem_PastUser {
    /**
     * @generated from protobuf field: bool isVip = 1;
     */
    isVip: boolean;
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.ReviewItem.VoteType
 */
export enum ReviewItem_VoteType {
    /**
     * @generated from protobuf enum value: None = 0;
     */
    None = 0,
    /**
     * @generated from protobuf enum value: ThumbsDown = -1;
     */
    ThumbsDown = -1,
    /**
     * @generated from protobuf enum value: ThumbsUp = 1;
     */
    ThumbsUp = 1
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.ReviewType
 */
export enum ReviewType {
    /**
     * @generated from protobuf enum value: NoneReviewType = 0;
     */
    NoneReviewType = 0,
    /**
     * @generated from protobuf enum value: NovelReviewType = 1;
     */
    NovelReviewType = 1,
    /**
     * @generated from protobuf enum value: ProductReviewType = 2;
     */
    ProductReviewType = 2
}
// ...
class SubmitReviewVoteRequest$Type extends MessageType<SubmitReviewVoteRequest> {
    constructor() {
        super("wuxiaworld.api.v2.SubmitReviewVoteRequest", [
            { no: 1, name: "id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "type", kind: "enum", T: () => ["wuxiaworld.api.v2.ReviewType", ReviewType] },
            { no: 3, name: "voteType", kind: "enum", T: () => ["wuxiaworld.api.v2.ReviewItem.VoteType", ReviewItem_VoteType], options: { "wuxiaworld.api.v2.required": { value: true }, "wuxiaworld.api.v2.display": { name: "Vote Type" } } }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.SubmitReviewVoteRequest
 */
export const SubmitReviewVoteRequest = new SubmitReviewVoteRequest$Type();
// ...
export const ReviewerSeriesInfo = new ReviewerSeriesInfo$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ReviewerEntityInfo$Type extends MessageType<ReviewerEntityInfo> {
    constructor() {
        super("wuxiaworld.api.v2.ReviewerEntityInfo", [
            { no: 1, name: "seriesInfo", kind: "message", oneof: "entityInfo", T: () => ReviewerSeriesInfo }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.ReviewerEntityInfo
 */
export const ReviewerEntityInfo = new ReviewerEntityInfo$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ReviewItem$Type extends MessageType<ReviewItem> {
    constructor() {
        super("wuxiaworld.api.v2.ReviewItem", [
            { no: 1, name: "id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "content", kind: "message", T: () => StringValue },
            { no: 3, name: "reviewedAt", kind: "message", T: () => Timestamp },
            { no: 4, name: "editedAt", kind: "message", T: () => Timestamp },
            { no: 5, name: "reviewUser", kind: "message", T: () => UserItem },
            { no: 6, name: "reviewUserPast", kind: "message", T: () => ReviewItem_PastUser },
            { no: 7, name: "votes", kind: "message", T: () => ReviewItem_Votes },
            { no: 8, name: "totalComments", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 9, name: "novelId", kind: "scalar", oneof: "type", T: 5 /*ScalarType.INT32*/ },
            { no: 10, name: "productId", kind: "scalar", oneof: "type", T: 5 /*ScalarType.INT32*/ },
            { no: 11, name: "voteType", kind: "enum", T: () => ["wuxiaworld.api.v2.ReviewItem.VoteType", ReviewItem_VoteType] },
            { no: 12, name: "hasPurchased", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 13, name: "rejectionInfo", kind: "message", T: () => ReviewItem_RejectionInfo },
            { no: 14, name: "reviewerEntityInfo", kind: "message", T: () => ReviewerEntityInfo }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.ReviewItem
 */
export const ReviewItem = new ReviewItem$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ReviewItem_Votes$Type extends MessageType<ReviewItem_Votes> {
    constructor() {
        super("wuxiaworld.api.v2.ReviewItem.Votes", [
            { no: 1, name: "upVotes", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "downVotes", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.ReviewItem.Votes
 */
export const ReviewItem_Votes = new ReviewItem_Votes$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ReviewItem_RejectionInfo$Type extends MessageType<ReviewItem_RejectionInfo> {
    constructor() {
        super("wuxiaworld.api.v2.ReviewItem.RejectionInfo", [
            { no: 1, name: "state", kind: "enum", T: () => ["wuxiaworld.api.v2.ReviewItem.RejectionInfo.State", ReviewItem_RejectionInfo_State] },
            { no: 2, name: "reason", kind: "message", T: () => StringValue },
            { no: 3, name: "rejectedAt", kind: "message", T: () => Timestamp }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.ReviewItem.RejectionInfo
 */
export const ReviewItem_RejectionInfo = new ReviewItem_RejectionInfo$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ReviewItem_PastUser$Type extends MessageType<ReviewItem_PastUser> {
    constructor() {
        super("wuxiaworld.api.v2.ReviewItem.PastUser", [
            { no: 1, name: "isVip", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.ReviewItem.PastUser
 */
export const ReviewItem_PastUser = new ReviewItem_PastUser$Type();